Add custom vocabulary to bag of words
A custom vocabulary can now be set when initialising or later, and this
is passed to underlying count-vectoriser class from `sci-kit learn`.

For large vocabularies, the count matrix will be very sparse, so when
converting to a data frame a sparse one is now used. Seaborn has
unfortunately not been updated to handle this format, so functionality
to remove absent words in a count matrix has been added and is used
before converting to a dense data frame.